+++ middlewares/validateBody.js
@@ -1,3 +1,5 @@
+const Joi = require("joi");
+
 const {HttpError} = require("../helpers");
 
 const validateBody = schema => {
@@ -12,4 +14,29 @@ const validateBody = schema => {
     return func;
 }
 
-module.exports = validateBody;
+const add = async (req, res) => {
+    const result = await contacts.addContact(req.body);
+    if (!result) {
+      res.status(400).json({"message": "missing required name field"})
+    }
+    res.status(201).json(result);
+}
+const updateById = async (req, res) => {
+    const { name, email, phone } = req.body;
+    const { id } = req.params;
+    if (!name && !email && !phone) {
+        res.status(400).json({"message": "missing fields"})
+    }
+    const result = await contacts.updateContact(id, req.body);
+    if (!result) {
+        return res.status(404).json ({
+          "message": "Not found"
+        });
+    }
+    res.status(200).json(result);
+}
+module.exports = {
+    add: ctrWrapper(add),
+    updateById: ctrWrapper(updateById),
+    validateBody,
+}
